PROTON-1400: WIP Use the mt broker example as the example instead of the previous st broker - The st broker didn't correctly respect the object access constraints from within handlers
