+++ examples/cpp/broker.cpp
@@ -1,5 +1,4 @@
 /*
- *
  * Licensed to the Apache Software Foundation (ASF) under one
  * or more contributor license agreements.  See the NOTICE file
  * distributed with this work for additional information
@@ -16,271 +15,300 @@
  * KIND, either express or implied.  See the License for the
  * specific language governing permissions and limitations
  * under the License.
- *
  */
 
 #include "options.hpp"
 
 #include <proton/connection.hpp>
+#include <proton/connection_options.hpp>
 #include <proton/container.hpp>
 #include <proton/default_container.hpp>
 #include <proton/delivery.hpp>
 #include <proton/error_condition.hpp>
+#include <proton/listen_handler.hpp>
 #include <proton/listener.hpp>
-#include <proton/messaging_handler.hpp>
 #include <proton/message.hpp>
-#include <proton/receiver_options.hpp>
-#include <proton/sender.hpp>
+#include <proton/messaging_handler.hpp>
 #include <proton/sender_options.hpp>
 #include <proton/source_options.hpp>
-#include <proton/target_options.hpp>
+#include <proton/target.hpp>
+#include <proton/thread_safe.hpp>
 #include <proton/tracker.hpp>
-#include <proton/transport.hpp>
-#include <proton/url.hpp>
 
+#include <atomic>
 #include <deque>
+#include <functional>
 #include <iostream>
-#include <list>
 #include <map>
+#include <mutex>
 #include <string>
 
 #include "fake_cpp11.hpp"
 
-/// A simple implementation of a queue.
+// Thread safe queue.
+// Stores messages, notifies subscribed connections when there is data.
 class queue {
   public:
-    queue(const std::string &name, bool dynamic = false) : name_(name), dynamic_(dynamic) {}
+    queue(const std::string& name) : name_(name) {}
 
     std::string name() const { return name_; }
 
-    void subscribe(proton::sender s) {
-        consumers_.push_back(s);
-    }
-
-    // Return true if queue can be deleted.
-    bool unsubscribe(proton::sender s) {
-        consumers_.remove(s);
-        return (consumers_.size() == 0 && (dynamic_ || messages_.size() == 0));
-    }
-
-    void publish(const proton::message &m) {
+    // Push a message onto the queue.
+    // If the queue was previously empty, notify subscribers it has messages.
+    // Called from receiver's connection.
+    void push(const proton::message &m) {
+        std::lock_guard<std::mutex> g(lock_);
         messages_.push_back(m);
-        dispatch(0);
+        if (messages_.size() == 1) { // Non-empty, notify subscribers
+            for (auto cb : callbacks_)
+                cb(this);
+            callbacks_.clear();
+        }
     }
 
-    void dispatch(proton::sender *s) {
-        while (deliver_to(s)) {}
-    }
-
-    bool deliver_to(proton::sender *s) {
-        // Deliver to single sender if supplied, else all consumers
-        int count = s ? 1 : consumers_.size();
-
-        if (!count) return false;
-
-        bool result = false;
-        sender_list::iterator it = consumers_.begin();
-
-        if (!s && count) {
-            s = &*it;
+    // If the queue is not empty, pop a message into m and return true.
+    // Otherwise save callback to be called when there are messages and return false.
+    // Called from sender's connection.
+    bool pop(proton::message& m, std::function<void(queue*)> callback) {
+        std::lock_guard<std::mutex> g(lock_);
+        if (messages_.empty()) {
+            callbacks_.push_back(callback);
+            return false;
+        } else {
+            m = std::move(messages_.front());
+            messages_.pop_front();
+            return true;
         }
-
-        while (messages_.size()) {
-            if (s->credit()) {
-                const proton::message& m = messages_.front();
-
-                s->send(m);
-                messages_.pop_front();
-                result = true;
-            }
-
-            if (--count) {
-                it++;
-            } else {
-                return result;
-            }
-        }
-
-        return false;
     }
 
   private:
-    typedef std::deque<proton::message> message_queue;
-    typedef std::list<proton::sender> sender_list;
-
-    std::string name_;
-    bool dynamic_;
-    message_queue messages_;
-    sender_list consumers_;
+    const std::string name_;
+    std::mutex lock_;
+    std::deque<proton::message> messages_;
+    std::vector<std::function<void(queue*)> > callbacks_;
 };
 
-/// A collection of queues and queue factory, used by a broker.
+/// Thread safe map of queues.
 class queues {
   public:
     queues() : next_id_(0) {}
-    virtual ~queues() {}
 
-    // Get or create a queue.
-    virtual queue &get(const std::string &address) {
-        if (address.empty()) {
-            throw std::runtime_error("empty queue name");
-        }
-
-        queue*& q = queues_[address];
-
-        if (!q) q = new queue(address);
-
-        return *q;
+    // Get or create the named queue.
+    queue* get(const std::string& name) {
+        std::lock_guard<std::mutex> g(lock_);
+        auto i = queues_.insert(queue_map::value_type(name, nullptr)).first;
+        if (!i->second)
+            i->second.reset(new queue(name));
+        return i->second.get();
     }
 
     // Create a dynamic queue with a unique name.
-    virtual queue &dynamic() {
+    queue* dynamic() {
         std::ostringstream os;
-        os << "q" << next_id_++;
-        queue *q = queues_[os.str()] = new queue(os.str(), true);
-
-        return *q;
+        os << "_dynamic_" << next_id_++;
+        return get(os.str());
     }
 
-    // Delete the named queue
-    virtual void erase(std::string &name) {
-        delete queues_[name];
-        queues_.erase(name);
-    }
+  private:
+    typedef std::map<std::string, std::unique_ptr<queue> > queue_map;
 
-  protected:
-    typedef std::map<std::string, queue *> queue_map;
+    std::mutex lock_;
     queue_map queues_;
-    int next_id_; // Use to generate unique queue IDs.
+    std::atomic<int> next_id_; // Use to generate unique queue IDs.
 };
 
-// A handler to implement broker logic
-class broker_handler : public proton::messaging_handler {
+/// Broker connection handler. Things to note:
+///
+/// 1. Each handler manages a single connection.
+///
+/// 2. For a *single connection* calls to proton::handler functions and calls to
+/// function objects passed to proton::event_loop::inject() are serialized,
+/// i.e. never called concurrently. Handlers can have per-connection state
+/// without needing locks.
+///
+/// 3. Handler/injected functions for *different connections* can be called
+/// concurrently. Resources used by multiple connections (e.g. the queues in
+/// this example) must be thread-safe.
+///
+/// 4. You can 'inject' work to be done sequentially using a connection's
+/// proton::event_loop. In this example, we create a std::function callback
+/// that we pass to queues, so they can notify us when they have messages.
+///
+class broker_connection_handler : public proton::messaging_handler {
   public:
-    broker_handler(queues& qs) : queues_(qs) {}
+    broker_connection_handler(queues& qs) : queues_(qs) {}
 
+    void on_connection_open(proton::connection& c) OVERRIDE {
+        // Create the has_messages callback for queue subscriptions.
+        //
+        // Make a std::shared_ptr to a thread_safe handle for our proton::connection.
+        // The connection's proton::event_loop will remain valid as a shared_ptr exists.
+        std::shared_ptr<proton::thread_safe<proton::connection> > ts_c = make_shared_thread_safe(c);
+
+        // Make a lambda function to inject a call to this->has_messages() via the proton::event_loop.
+        // The function is bound to a shared_ptr so this is safe. If the connection has already closed
+        // proton::event_loop::inject() will drop the callback.
+        has_messages_callback_ = [this, ts_c](queue* q) mutable {
+            ts_c->event_loop().inject(
+                std::bind(&broker_connection_handler::has_messages, this, q));
+        };
+
+        c.open();            // Accept the connection
+    }
+
+    // A sender sends messages from a queue to a subscriber.
     void on_sender_open(proton::sender &sender) OVERRIDE {
-        proton::source src(sender.source());
-        queue *q;
-        if (src.dynamic()) {
-            q = &queues_.dynamic();
-        } else if (!src.address().empty()) {
-            q = &queues_.get(src.address());
-        } else {
-            sender.close(proton::error_condition("No queue address supplied"));
-            return;
-        }
-        sender.open(proton::sender_options().source(proton::source_options().address(q->name())));
-        q->subscribe(sender);
-        std::cout << "broker outgoing link from " << q->name() << std::endl;
+        queue *q = sender.source().dynamic() ?
+            queues_.dynamic() : queues_.get(sender.source().address());
+        sender.open(proton::sender_options().source((proton::source_options().address(q->name()))));
+        std::cout << "sending from " << q->name() << std::endl;
     }
 
-    void on_receiver_open(proton::receiver &receiver) OVERRIDE {
-        std::string address = receiver.target().address();
-        if (!address.empty()) {
-            receiver.open(proton::receiver_options().target(proton::target_options().address(address)));
-            std::cout << "broker incoming link to " << address << std::endl;
+    // We have credit to send a message.
+    void on_sendable(proton::sender &s) OVERRIDE {
+        queue* q = sender_queue(s);
+        if (!do_send(q, s))     // Queue is empty, save ourselves in the blocked set.
+            blocked_.insert(std::make_pair(q, s));
+    }
+
+    // A receiver receives messages from a publisher to a queue.
+    void on_receiver_open(proton::receiver &r) OVERRIDE {
+        std::string qname = r.target().address();
+        if (qname == "shutdown") {
+            std::cout << "broker shutting down" << std::endl;
+            // Sending to the special "shutdown" queue stops the broker.
+            r.connection().container().stop(
+                proton::error_condition("shutdown", "stop broker"));
         } else {
-            receiver.close(proton::error_condition("No queue address supplied"));
+            std::cout << "receiving to " << qname << std::endl;
         }
     }
 
-    void unsubscribe(proton::sender lnk) {
-        std::string address = lnk.source().address();
+    // A message is received.
+    void on_message(proton::delivery &d, proton::message &m) OVERRIDE {
+        std::string qname = d.receiver().target().address();
+        queues_.get(qname)->push(m);
+    }
 
-        if (queues_.get(address).unsubscribe(lnk)) {
-            queues_.erase(address);
-        }
+    void on_session_close(proton::session &session) OVERRIDE {
+        // Erase all blocked senders that belong to session.
+        auto predicate = [session](const proton::sender& s) {
+            return s.session() == session;
+        };
+        erase_sender_if(blocked_.begin(), blocked_.end(), predicate);
     }
 
     void on_sender_close(proton::sender &sender) OVERRIDE {
-        unsubscribe(sender);
+        // Erase sender from the blocked set.
+        auto range = blocked_.equal_range(sender_queue(sender));
+        auto predicate = [sender](const proton::sender& s) { return s == sender; };
+        erase_sender_if(range.first, range.second, predicate);
     }
 
-    void on_connection_close(proton::connection &c) OVERRIDE {
-        remove_stale_consumers(c);
+    void on_error(const proton::error_condition& e) OVERRIDE {
+        std::cerr << "error: " << e.what() << std::endl;
+    }
+    // The container calls on_transport_close() last.
+    void on_transport_close(proton::transport&) OVERRIDE {
+        delete this;            // All done.
     }
 
-    void on_transport_close(proton::transport &t) OVERRIDE {
-        remove_stale_consumers(t.connection());
+  private:
+    typedef std::multimap<queue*, proton::sender> blocked_map;
+
+    // Get the queue associated with a sender.
+    queue* sender_queue(const proton::sender& s) {
+        return queues_.get(s.source().address()); // Thread safe.
     }
 
-    void on_transport_error(proton::transport &t) OVERRIDE {
-        std::cout << "broker client disconnect: " << t.error().what() << std::endl;
+    // Only called if we have credit. Return true if we sent a message.
+    bool do_send(queue* q, proton::sender &s) {
+        proton::message m;
+        bool popped =  q->pop(m, has_messages_callback_);
+        if (popped)
+            s.send(m);
+        /// if !popped the queue has saved the callback for later.
+        return popped;
     }
 
-    void on_error(const proton::error_condition &c) OVERRIDE {
-        std::cerr << "broker error: " << c.what() << std::endl;
-    }
-
-    void remove_stale_consumers(proton::connection connection) {
-        proton::sender_range r = connection.senders();
-        for (proton::sender_iterator i = r.begin(); i != r.end(); ++i) {
-            if (i->active())
-                unsubscribe(*i);
+    // Called via the connection's proton::event_loop when q has messages.
+    // Try all the blocked senders.
+    void has_messages(queue* q) {
+        auto range = blocked_.equal_range(q);
+        for (auto i = range.first; i != range.second;) {
+            if (i->second.credit() <= 0 || do_send(q, i->second))
+                i = blocked_.erase(i); // No credit or send was successful, stop blocked.
+            else
+                ++i;            // have credit, didn't send, keep blocked
         }
     }
 
-    void on_sendable(proton::sender &s) OVERRIDE {
-        std::string address = s.source().address();
-
-        queues_.get(address).dispatch(&s);
+    // Use to erase closed senders from blocked_ set.
+    template <class Predicate>
+    void erase_sender_if(blocked_map::iterator begin, blocked_map::iterator end, Predicate p) {
+        for (auto i = begin; i != end; ) {
+            if (p(i->second))
+                i = blocked_.erase(i);
+            else
+                ++i;
+        }
     }
 
-    void on_message(proton::delivery &d, proton::message &m) OVERRIDE {
-        std::string address = d.receiver().target().address();
-        queues_.get(address).publish(m);
-    }
-
-  protected:
     queues& queues_;
+    blocked_map blocked_;
+    std::function<void(queue*)> has_messages_callback_;
+    proton::connection connection_;
 };
 
 
-// The broker
 class broker {
   public:
-    broker(const std::string& url) : handler_(url, queues_) {}
+    broker(const std::string addr) :
+        container_("mt_broker"), listener_(queues_)
+    {
+        container_.listen(addr, listener_);
+        std::cout << "broker listening on " << addr << std::endl;
+    }
 
-    proton::messaging_handler& handler() { return handler_; }
+    void run() {
+        container_.run(/* std::thread::hardware_concurrency() */);
+    }
 
   private:
-    class my_handler : public broker_handler {
-      public:
-        my_handler(const std::string& u, queues& qs) : broker_handler(qs), url_(u) {}
+    struct listener : public proton::listen_handler {
+        listener(queues& qs) : queues_(qs) {}
 
-        void on_container_start(proton::container &c) OVERRIDE {
-            c.listen(url_);
-            std::cout << "broker listening on " << url_ << std::endl;
+        proton::connection_options on_accept(proton::listener&) OVERRIDE{
+            return proton::connection_options().handler(*(new broker_connection_handler(queues_)));
         }
 
-      private:
-        const std::string& url_;
+        void on_error(proton::listener&, const std::string& s) OVERRIDE {
+            std::cerr << "listen error: " << s << std::endl;
+            throw std::runtime_error(s);
+        }
+        queues& queues_;
     };
 
-  private:
     queues queues_;
-    my_handler handler_;
+    proton::container container_;
+    listener listener_;
 };
 
 int main(int argc, char **argv) {
-    std::string url("0.0.0.0");
+    // Command line options
+    std::string address("0.0.0.0");
     example::options opts(argc, argv);
 
-    opts.add_value(url, 'a', "address", "listen on URL", "URL");
+    opts.add_value(address, 'a', "address", "listen on URL", "URL");
 
     try {
         opts.parse();
-
-        broker b(url);
-        proton::default_container(b.handler()).run();
-
+        broker(address).run();
         return 0;
     } catch (const example::bad_option& e) {
         std::cout << opts << std::endl << e.what() << std::endl;
     } catch (const std::exception& e) {
-        std::cerr << e.what() << std::endl;
+        std::cerr << "broker shutdown: " << e.what() << std::endl;
     }
-
     return 1;
 }
